+++ src/Airbnb.AppService/Commands/Listing/CreateListing/CreateListingCommand.cs
@@ -1,8 +1,10 @@
-﻿using Airbnb.Core.Commands;
+﻿using Airbnb.AppService.Responses.Listing;
+using Airbnb.Core.Commands;
+using Airbnb.SharedKernel;
 
 namespace Airbnb.AppService.Commands.Listing.CreateListing;
 
-public sealed class CreateListingCommand : ICommand
+public sealed class CreateListingCommand : ICommand<Result<CreateListingResponse>>
 {
 
 }