+++ src/Airbnb.AppService/Responses/Listing/CreateListingResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace Airbnb.AppService.Responses.Listing;
+
+public class CreateListingResponse
+{
+    
+}