+++ src/Airbnb.Infra.Broker/Extensions.cs
@@ -1,0 +1,13 @@
+﻿using Airbnb.Infra.Broker.Publisher;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Airbnb.Infra.Broker;
+
+public static class Extensions
+{
+    public static IServiceCollection AddBroker(this IServiceCollection services)
+    {
+        services.AddScoped<IEventPublisher, ServiceBusPublisher>();
+        return services;
+    }
+}