+++ src/Airbnb.Infra.DependencyInjection/DependencyInjection.cs
@@ -1,6 +1,9 @@
 ﻿using Airbnb.AppService;
 using Airbnb.Core;
+using Airbnb.Infra.Broker;
+using Airbnb.Infra.Broker.Publisher;
 using Airbnb.Infra.Repository;
+using Airbnb.SharedKernel.Audit;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -13,6 +16,13 @@ public static class DependencyInjection
         services.AddAppService();
         services.AddCore();
         services.AddRepository();
+        services.AddBroker();
+
+        services.AddSingleton<AuditPublisherDelegate>(sp =>
+        {
+            var publisher = sp.GetRequiredService<IEventPublisher>();
+            return message => publisher.PublishAsync(message, "audit-requests");
+        });
 
         return services;
     }