+++ src/Airbnb.SharedKernel/AirbnbSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace Airbnb.SharedKernel;
+
+public class AirbnbSettings
+{
+    public const string ServiceName = "Airbnb.Services.Listing";
+}