+++ src/Airbnb.SharedKernel/Error.cs
@@ -1,0 +1,13 @@
+﻿namespace Airbnb.SharedKernel;
+
+[Serializable]
+public class Error(string code, string message)
+{
+    public Error()
+        : this(string.Empty, string.Empty)
+    {
+    }
+
+    public string Code { get; } = code;
+    public string Message { get; } = message;
+}