+++ src/Airbnb.SharedKernel/ErrorCode.cs
@@ -1,0 +1,6 @@
+﻿namespace Airbnb.SharedKernel;
+
+public static class ErrorCode
+{
+    public const string FailedToCreateListing = "FailedToCreateListing";
+}